# swagger.yaml
openapi: 3.0.3
info:
  title: Equipment Loan Backend API
  version: 1.0.0
  description: Expressで実装された備品貸出API
servers:
  - url: http://localhost:4000
    description: Local development
paths:
...
  /api/equipment:
    get:
      tags:
        - Equipment
      summary: 備品一覧取得
      description: member以上のbearerAuthが必要。条件絞り込みとページングに対応し、ID昇順で一覧を返す
      operationId: getEquipment
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - available
              - in_use
              - reserved
        - name: category
          in: query
          required: false
          description: ノートPC / モニタ / 周辺機器 のいずれか
          schema:
            type: string
            enum:
              - ノートPC
              - モニタ
              - 周辺機器
        - name: location
          in: query
          required: false
          description: 部分一致検索（空文字不可）
          schema:
            type: string
            minLength: 1
        - name: keyword
          in: query
          required: false
          description: id / name / note に対する部分一致検索（空文字不可）
          schema:
            type: string
            minLength: 1
        - name: page
          in: query
          required: false
          description: limit とセットで指定
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          required: false
          description: page とセットで指定（1-100）
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: 備品一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EquipmentItem"
              examples:
                list:
                  value:
                    - id: eq-001
                      name: ThinkPad X1 Carbon Gen 12
                      category: ノートPC
                      location: 東京オフィス 4F
                      status: available
                      assignee: null
                      requestedBy: null
                      reservationPeriod: 即日 - 2026/05/12
                      specs:
                        - Core Ultra 7
                        - 32GB RAM
                        - 14 inch
                      note: 営業同行用に軽量モデル。USB-C 充電器を付属。
        "403":
          description: member未満、未ログイン、または不正トークン
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                forbidden:
                  value:
                    message: forbidden
        "400":
          description: クエリ不正
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Equipment
      summary: 備品追加
      description: admin の bearerAuth が必要。新規備品を登録する
      operationId: createEquipment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentUpsertRequest"
      responses:
        "201":
          description: 作成成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentItem"
        "400":
          description: 入力不正
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: admin未満、未ログイン、または不正トークン
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: 同一IDが既に存在
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/equipment/{id}:
    get:
      tags:
        - Equipment
      summary: 備品詳細取得
      description: member以上のbearerAuthが必要。備品IDで詳細を返す
      operationId: getEquipmentById
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: "備品ID（例: eq-001）"
          schema:
            type: string
      responses:
        "200":
          description: 備品詳細
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentItem"
        "403":
          description: member未満、未ログイン、または不正トークン
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                forbidden:
                  value:
                    message: forbidden
        "404":
          description: 備品が見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                notFound:
                  value:
                    message: equipment not found
    patch:
      tags:
        - Equipment
      summary: 備品更新
      description: admin の bearerAuth が必要。指定IDの備品情報を部分更新する
      operationId: updateEquipment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentPatchRequest"
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentItem"
        "400":
          description: 入力不正
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: admin未満、未ログイン、または不正トークン
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 備品が見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Equipment
      summary: 備品削除
      description: admin の bearerAuth が必要。指定IDの備品を削除する
      operationId: deleteEquipment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: 削除成功
        "403":
          description: admin未満、未ログイン、または不正トークン
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 備品が見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/equipment/{id}/checkout:
    post:
      tags:
        - Equipment
      summary: 備品貸出開始
      description: admin の bearerAuth が必要。空き備品を貸出中へ変更する
      operationId: checkoutEquipment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentCheckoutRequest"
      responses:
        "200":
          description: 貸出開始成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentItem"
        "400":
          description: 入力不正
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: admin未満、未ログイン、または不正トークン
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 備品が見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: 備品が貸出可能状態でない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/equipment/{id}/return:
    post:
      tags:
        - Equipment
      summary: 備品返却処理
      description: admin の bearerAuth が必要。貸出中/予約中備品を空きへ戻す
      operationId: returnEquipment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 返却処理成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentItem"
        "403":
          description: admin未満、未ログイン、または不正トークン
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 備品が見つからない
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: 既に空き状態
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/active-loans:
    get:
      tags:
        - Equipment
      summary: 貸出中または予約中の備品一覧取得
      description: member以上のbearerAuthが必要。status が available 以外の備品を返す
      operationId: getActiveLoans
      security:
        - bearerAuth: []
      responses:
        "403":
          description: member未満、未ログイン、または不正トークン
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                forbidden:
                  value:
                    message: forbidden
        "200":
          description: 貸出中・予約中備品一覧
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EquipmentItem"
              examples:
                active:
                  value:
                    - id: eq-002
                      name: Dell 27 4K USB-C Monitor
                      category: モニタ
                      location: 大阪オフィス 2F
                      status: in_use
                      assignee: 佐藤 亮
                      requestedBy: 佐藤 亮
                      reservationPeriod: 2026/05/01 - 2026/05/15
                      specs:
                        - 27 inch
                        - 4K
                        - USB-C 90W
                      note: デザインレビュー室で利用中。返却時は高さ調整台も回収。
                    - id: eq-003
                      name: MacBook Pro 14 M4
                      category: ノートPC
                      location: 福岡オフィス 3F
                      status: reserved
                      assignee: null
                      requestedBy: 山本 愛
                      reservationPeriod: 2026/05/10 - 2026/05/20
                      specs:
                        - M4 Pro
                        - 24GB RAM
                        - 1TB SSD
                      note: 展示会向けの動画編集用。貸出開始前にFinal Cutの起動確認が必要。
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EquipmentItem:
      type: object
      required:
        - id
        - name
        - category
        - location
        - status
        - assignee
        - requestedBy
        - reservationPeriod
        - specs
        - note
      properties:
        id:
          type: string
          example: eq-001
        name:
          type: string
          example: ThinkPad X1 Carbon Gen 12
        category:
          type: string
          example: ノートPC
        location:
          type: string
          example: 東京オフィス 4F
        status:
          type: string
          enum:
            - available
            - in_use
            - reserved
          example: available
        assignee:
          type: string
          nullable: true
          example: null
        requestedBy:
          type: string
          nullable: true
          example: null
        reservationPeriod:
          type: string
          example: 即日 - 2026/05/12
        specs:
          type: array
          items:
            type: string
          example:
            - Core Ultra 7
            - 32GB RAM
            - 14 inch
        note:
          type: string
          example: 営業同行用に軽量モデル。USB-C 充電器を付属。
    EquipmentUpsertRequest:
      type: object
      required:
        - id
        - name
        - category
        - location
        - reservationPeriod
        - specs
        - note
      properties:
        id:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        category:
          type: string
          enum:
            - ノートPC
            - モニタ
            - 周辺機器
        location:
          type: string
          minLength: 1
        status:
          type: string
          enum:
            - available
            - in_use
            - reserved
        assignee:
          type: string
          nullable: true
        requestedBy:
          type: string
          nullable: true
        reservationPeriod:
          type: string
          pattern: ".*\\d{4}\\/\\d{2}\\/\\d{2}$"
        specs:
          type: array
          minItems: 1
          items:
            type: string
            minLength: 1
        note:
          type: string
          minLength: 1
    EquipmentPatchRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
          enum:
            - ノートPC
            - モニタ
            - 周辺機器
        location:
          type: string
          minLength: 1
        status:
          type: string
          enum:
            - available
            - in_use
            - reserved
        assignee:
          type: string
          nullable: true
        requestedBy:
          type: string
          nullable: true
        reservationPeriod:
          type: string
          pattern: ".*\\d{4}\\/\\d{2}\\/\\d{2}$"
        specs:
          type: array
          items:
            type: string
            minLength: 1
        note:
          type: string
          minLength: 1
    EquipmentCheckoutRequest:
      type: object
      required:
        - assignee
        - requestedBy
        - reservationPeriod
      properties:
        assignee:
          type: string
          minLength: 1
        requestedBy:
          type: string
          minLength: 1
        reservationPeriod:
          type: string
          pattern: ".*\\d{4}\\/\\d{2}\\/\\d{2}$"
...
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: equipment not found